last few feedback adjustments
